frontend/requirements.txt
streamlit
requests
pillow
pyyaml
pytesseract
speechrecognition
langdetect